backend/requirements/dev.txt
-r base.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
bandit==1.7.5

# Pre-commit hooks
pre-commit==3.6.0

# Development Server
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Database Tools
django-extensions==3.2.3

# API Testing
pytest-asyncio==0.21.1
httpx==0.25.2 